# docs/api/openapi-snapshot.json
{
  "components" : {
    "schemas" : {
...
      "MediaAcceptDto" : {
        "properties" : {
          "assetId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "bearerFormat" : "JWT",
        "description" : "JWT token from Keycloak. Must contain tenant claim (`tenant_id`, `tenantId`, or `tid`).",
        "scheme" : "bearer",
        "type" : "http"
      },
      "catalog-scopes" : {
        "description" : "Least-privilege machine/integration access via OAuth2 client-credentials scopes (issue #206 [AI-4]). A catalog:read-only token lists products (200) but cannot mutate them (403).",
        "flows" : {
          "clientCredentials" : {
            "scopes" : {
              "catalog:read" : "Read the product catalog (list/get products)",
              "catalog:write" : "Create, update, and delete products and product images",
              "customers:read" : "Reserved for the MCP model — defined-only, not enforced this phase",
              "customers:write" : "Enforced; gates POST /customers (Phase 25 [AI-02])",
              "orders:read" : "Reserved for the MCP model — defined, not yet enforced",
              "orders:write" : "Enforced; gates POST /orders (Phase 25 [AI-02])"
            },
            "tokenUrl" : "http://localhost:8085/realms/jtoye-dev/protocol/openid-connect/token"
          }
        },
        "type" : "oauth2"
      },
      "tenant-header" : {
        "description" : "Dev fallback: UUID of tenant (only used when JWT lacks tenant claim)",
        "in" : "header",
        "name" : "X-Tenant-Id",
        "type" : "apiKey"
      }
    }
  },
  "info" : {
    "contact" : {
      "email" : "engineering@jtoye.uk",
      "name" : "J'Toye Engineering"
    },
...
    "license" : {
      "name" : "Proprietary",
      "url" : "https://jtoye.uk/license"
    },
    "title" : "J'Toye OaaS Core API",
    "version" : "1.2.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/products" : {
      "get" : {
        "description" : "Returns a paginated list of products for the authenticated tenant. All products include mandatory allergen and ingredient information per Natasha's Law. Optional shopId query param narrows to one shop of the tenant, server-side (requires at least STAFF on that shop).",
        "operationId" : "list_3",
        "parameters" : [
          {
            "description" : "Optional shop to narrow the list to",
            "in" : "query",
            "name" : "shopId",
            "required" : false,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageProductDto"
                }
              }
            },
            "description" : "Successfully retrieved products"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageProductDto"
                }
              }
            },
            "description" : "Unauthorized - missing or invalid JWT"
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageProductDto"
                }
              }
            },
            "description" : "Caller has no grant on the requested shopId"
          }
        },
        "security" : [
          {
            "tenant-header" : [ ]
          },
          {
            "bearer-jwt" : [ ]
          }
        ],
        "summary" : "List products",
        "tags" : [
          "Products"
        ]
      },
      "post" : {
        "description" : "Creates a new product. Requires ingredients_text, allergen_mask, and price per Natasha's Law (UK) and business requirements.",
        "operationId" : "create_3",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateProductRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductDto"
                }
              }
            },
            "description" : "Product created successfully"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductDto"
                }
              }
            },
            "description" : "Validation error - missing required fields or invalid price"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductDto"
                }
              }
            },
            "description" : "Product SKU already exists for this tenant"
          }
        },
        "security" : [
          {
            "tenant-header" : [ ]
          },
          {
            "bearer-jwt" : [ ]
          }
        ],
        "summary" : "Create product",
        "tags" : [
          "Products"
        ]
      }
    },
    "/api/v1/products/bulk/csv" : {
      "post" : {
        "description" : "Import multiple products from a CSV file. Returns created products and per-row errors.",
        "operationId" : "bulkImportCsv",
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "properties" : {
                  "file" : {
                    "format" : "binary",
                    "type" : "string"
                  }
                },
                "required" : [
                  "file"
                ],
                "type" : "object"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BulkImportResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [
          {
            "tenant-header" : [ ]
          },
          {
            "bearer-jwt" : [ ]
          }
        ],
        "summary" : "Bulk import from CSV",
        "tags" : [
          "Products"
        ]
      }
    },
    "/api/v1/products/bulk/images" : {
      "post" : {
        "description" : "Upload multiple food photos. AI identifies each item and creates draft products.",
        "operationId" : "bulkImportImages",
        "parameters" : [
          {
            "in" : "query",
            "name" : "files",
            "required" : true,
            "schema" : {
              "items" : {
                "format" : "binary",
                "type" : "string"
              },
              "type" : "array"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BulkImportResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [
          {
            "tenant-header" : [ ]
          },
          {
            "bearer-jwt" : [ ]
          }
        ],
        "summary" : "Bulk import from images",
        "tags" : [
          "Products"
        ]
      }
    },
    "/api/v1/products/search" : {
      "get" : {
        "description" : "Full-text search (GIN/tsvector, per-word prefix matching, relevance-ranked) over title, category, description, ingredients and dietary tags, plus SKU prefix lookup. Returns a JSON array (wire contract frozen: edge WhatsApp flow and dashboard both consume it) of at most one page; optional page/size params paginate, with size capped by the global pageable maximum (100). Optional shopId narrows to one shop of the tenant (requires at least STAFF on that shop).",
        "operationId" : "search_1",
        "parameters" : [
          {
            "in" : "query",
            "name" : "q",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "Optional shop to narrow the search to",
            "in" : "query",
            "name" : "shopId",
            "required" : false,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ProductDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [
          {
            "tenant-header" : [ ]
          },
          {
            "bearer-jwt" : [ ]
          }
        ],
        "summary" : "Search products",
        "tags" : [
          "Products"
        ]
      }
    },
    "/api/v1/products/template" : {
      "get" : {
        "description" : "Returns a CSV template file for bulk product import",
        "operationId" : "downloadTemplate",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "format" : "byte",
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [
          {
            "tenant-header" : [ ]
          },
          {
            "bearer-jwt" : [ ]
          }
        ],
        "summary" : "Download CSV template",
        "tags" : [
          "Products"
        ]
      }
    },
    "/api/v1/products/{id}" : {
      "delete" : {
        "description" : "Deletes a product for the authenticated tenant",
        "operationId" : "delete_2",
        "parameters" : [
          {
            "description" : "Product ID",
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "204" : {
            "description" : "Product deleted successfully"
          },
          "404" : {
            "description" : "Product not found"
          }
        },
        "security" : [
          {
            "tenant-header" : [ ]
          },
          {
            "bearer-jwt" : [ ]
          }
        ],
        "summary" : "Delete product",
        "tags" : [
          "Products"
        ]
      },
      "get" : {
        "description" : "Returns a single product by ID for the authenticated tenant",
        "operationId" : "getById_2",
        "parameters" : [
          {
            "description" : "Product ID",
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductDto"
                }
              }
            },
            "description" : "Product found"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductDto"
                }
              }
            },
            "description" : "Product not found"
          }
        },
        "security" : [
          {
            "tenant-header" : [ ]
          },
          {
            "bearer-jwt" : [ ]
          }
        ],
        "summary" : "Get product by ID",
        "tags" : [
          "Products"
        ]
      },
      "put" : {
        "description" : "Updates an existing product for the authenticated tenant",
        "operationId" : "update_2",
        "parameters" : [
          {
            "description" : "Product ID",
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateProductRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductDto"
                }
              }
            },
            "description" : "Product updated successfully"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductDto"
                }
              }
            },
            "description" : "Validation error"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductDto"
                }
              }
            },
            "description" : "Product not found"
          }
        },
        "security" : [
          {
            "tenant-header" : [ ]
          },
          {
            "bearer-jwt" : [ ]
          }
        ],
        "summary" : "Update product",
        "tags" : [
          "Products"
        ]
      }
    },
    "/api/v1/products/{id}/image" : {
      "delete" : {
        "description" : "Removes the image from a product",
        "operationId" : "removeImage",
        "parameters" : [
          {
            "description" : "Product ID",
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductDto"
                }
              }
            },
            "description" : "Image removed"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductDto"
                }
              }
            },
            "description" : "Product not found"
          }
        },
        "security" : [
          {
            "tenant-header" : [ ]
          },
          {
            "bearer-jwt" : [ ]
          }
        ],
        "summary" : "Remove product image",
        "tags" : [
          "Products"
        ]
      },
      "post" : {
        "description" : "Quarantines the raw upload and queues it for magic-byte sniffing, decode-verify, EXIF strip and WebP normalization. Returns 202 with the media_asset id; the client polls the product to observe PENDING -> ACTIVE/FAILED. Carries the uniform Idempotency-Key contract; oversize/replay errors are RFC 7807 typed.",
        "operationId" : "accept",
        "parameters" : [
          {
            "description" : "Product ID",
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          },
          {
            "in" : "header",
            "name" : "Content-Length",
            "required" : false,
            "schema" : {
              "format" : "int64",
              "type" : "integer"
            }
          },
          {
            "in" : "header",
            "name" : "Idempotency-Key",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "in" : "query",
            "name" : "is_primary",
            "required" : false,
            "schema" : {
              "default" : true,
              "type" : "boolean"
            }
          },
          {
            "in" : "query",
            "name" : "sort_order",
            "required" : false,
            "schema" : {
              "default" : 0,
              "format" : "int32",
              "type" : "integer"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "properties" : {
                  "file" : {
                    "format" : "binary",
                    "type" : "string"
                  }
                },
                "required" : [
                  "file"
                ],
                "type" : "object"
              }
            }
          }
        },
        "responses" : {
          "202" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MediaAcceptDto"
                }
              }
            },
            "description" : "Accepted — upload quarantined, normalization queued"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MediaAcceptDto"
                }
              }
            },
            "description" : "Missing Idempotency-Key or malformed request"
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MediaAcceptDto"
                }
              }
            },
            "description" : "Missing catalog:write scope or shop-role access"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MediaAcceptDto"
                }
              }
            },
            "description" : "Product not found"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MediaAcceptDto"
                }
              }
            },
            "description" : "Idempotency-Key request already in progress"
          },
          "413" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MediaAcceptDto"
                }
              }
            },
            "description" : "Upload exceeds the maximum permitted size"
          },
          "422" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MediaAcceptDto"
                }
              }
            },
            "description" : "Idempotency-Key reused with a different body"
          }
        },
        "security" : [
          {
            "tenant-header" : [ ]
          },
          {
            "bearer-jwt" : [ ]
          }
        ],
        "summary" : "Upload a product image (async accept)",
        "tags" : [
          "Product Media"
        ]
      }
    },
    "/api/v1/products/{id}/image/analyze" : {
      "post" : {
        "description" : "Identifies the food item, suggests ingredients, category, and dietary info without saving the image.",
        "operationId" : "analyzeImage",
        "parameters" : [
          {
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "properties" : {
                  "file" : {
                    "format" : "binary",
                    "type" : "string"
                  }
                },
                "required" : [
                  "file"
                ],
                "type" : "object"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ImageAnalysisResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [
          {
            "tenant-header" : [ ]
          },
          {
            "bearer-jwt" : [ ]
          }
        ],
        "summary" : "Analyze product image with AI",
        "tags" : [
          "Products"
        ]
      }
    },
...
  },
  "security" : [
    {
      "bearer-jwt" : [ ]
    },
    {
      "tenant-header" : [ ]
    }
  ],
...
}